updating scripts & templates

+++ mountdisk.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
-mkfs.ext4 $1
+if [ $3 == "format" ]
+then
+	mkfs.ext4 $1
+fi
 mkdir $2
 mount $1 $2
